decoder block done
